requirements.txt
markovify
mecab-python3
neologdn
requests
requests-oauthlib
